requirements.txt
# Core dependencies
websockets>=11.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=1.4.0,<2.0.0
aiofiles>=23.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
httpx>=0.27.0
pymysql>=1.1.0
cryptography>=41.0.0
psutil>=5.9.0
schedule>=1.2.0

# Scientific/ML dependencies with flexible versions
numpy>=1.21.0,<1.27.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# OpenAI API
openai>=1.3.0

# Note: ta-lib removed as it's difficult to install and not essential
# asyncio is built-in to Python, no need to install
